% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyclopsInBeast.R
\docType{package}
\name{CyclopsInBeast-package}
\alias{CyclopsInBeast}
\alias{CyclopsInBeast-package}
\title{CyclopsInBeast: Cyclops for Fast Likelihood and Gradient Evaluations in BEAST}
\description{
Routines for exploiting Cyclops in Bayesian analyses through the BEAST library
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/CyclopsInBeast}
  \item Report bugs at \url{https://github.com/OHDSI/CyclopsInBeast/issues}
}

}
\author{
\strong{Maintainer}: Marc A. Suchard \email{msuchard@ucla.edu}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Fan Bu fanbu@g.ucla.edu
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
